refactor: move default props
